dom/workers/test/test_bug1014466.html

<!--
2   Any copyright is dedicated to the Public Domain.
3   http://creativecommons.org/publicdomain/zero/1.0/
4 -->
<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1014466
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1014466</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1014466">Mozilla Bug 1014466</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script type="application/javascript">

  var worker = new Worker("bug1014466_worker.js");

  worker.onmessage = function(event) {
    if (event.data.type == 'finish') {
      SimpleTest.finish();
    } else if (event.data.type == 'status') {
      ok(event.data.status, event.data.msg);
    }
  };

  worker.postMessage(true);

  SimpleTest.waitForExplicitFinish();

</script>
</pre>
</body>
</html>
